+++ app/src/main/AndroidManifest.xml
@@ -1,26 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.zoonref.demo"
-          xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest
+    package="com.zoonref.demo"
+    xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
-        android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
             android:name=".DemoActivity"
-            android:label=""
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
-
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
         <activity
             android:name=".ProfileActivity"
-            android:label=""
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
     </application>
