+++ library/src/main/AndroidManifest.xml
@@ -1,12 +1,1 @@
-<manifest package="com.zoonref"
-          xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application
-        android:allowBackup="true"
-        android:label="@string/app_name"
-        android:supportsRtl="true"
-        >
-
-    </application>
-
-</manifest>
+<manifest package="com.zoonref.viewbehavior"/>
