+++ library/src/test/java/com/zoonref/viewbehavior/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.zoonref.zoonref;
+package com.zoonref.simpleviewbehavior;
 
 import org.junit.Test;
